Add Read.getObjectType
Problem: sometimes you have an object ID but need to know what type of
object the ID points at in order to determine what to do with it but
there is no way of getting this information from automerge.

Solution: add Read.getObjectType which returns the type of the object
ID, or `Optional.empty` if the object is not in the document.