Check for null object IDs
Problem: at various places in the rust codebase we receive an object ID
from the java side in the form of a `jobject` which is a pointer to a
`ObjectId`. This pointer can be null but we were not checking whether it
was null and consequently were panicking and crashing the entire JVM.

Solution: check for null when loading object IDs and throw an
IllegalArgumentException if the `ObjectId` is null. This required a
macro to avoid lots of code handling the early return.